backend/app/requirement.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
tensorflow==2.13.0
numpy==1.24.3
pillow==10.0.1
opencv-python==4.8.1.78
pydantic==2.5.0